+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/pages/userregistration/UserRegistrationApiService.kt
@@ -3,10 +3,11 @@ package de.sovity.authorityportal.web.pages.userregistration
 import de.sovity.authorityportal.api.model.CreateOrganizationRequest
 import de.sovity.authorityportal.api.model.IdResponse
 import de.sovity.authorityportal.api.model.UserRegistrationStatusResult
+import de.sovity.authorityportal.db.jooq.enums.OrganizationRegistrationStatus
 import de.sovity.authorityportal.db.jooq.enums.UserRegistrationStatus
+import de.sovity.authorityportal.web.pages.organizationmanagement.OrganizationRegistrationApiService
 import de.sovity.authorityportal.web.services.OrganizationService
 import de.sovity.authorityportal.web.services.UserService
-import de.sovity.authorityportal.web.pages.organizationmanagement.OrganizationManagementApiService
 import de.sovity.authorityportal.web.thirdparty.keycloak.KeycloakService
 import de.sovity.authorityportal.web.thirdparty.keycloak.model.OrganizationRole
 import de.sovity.authorityportal.web.utils.idmanagement.MdsIdUtils
@@ -21,7 +22,7 @@ class UserRegistrationApiService {
     lateinit var keycloakService: KeycloakService
 
     @Inject
-    lateinit var organizationManagementApiService: OrganizationManagementApiService
+    lateinit var organizationRegistrationApiService: OrganizationRegistrationApiService
 
     @Inject
     lateinit var organizationService: OrganizationService
@@ -48,7 +49,7 @@ class UserRegistrationApiService {
         keycloakService.joinOrganization(userId, mdsId, OrganizationRole.PARTICIPANT_ADMIN)
         keycloakService.forceLogout(userId)
 
-        organizationService.createOrganization(userId, mdsId, organization)
+        organizationService.createOrganization(userId, mdsId, organization, OrganizationRegistrationStatus.PENDING)
         val user = userService.getUserOrThrow(userId)
         val firstUser = user.registrationStatus == UserRegistrationStatus.FIRST_USER
         user.registrationStatus = UserRegistrationStatus.PENDING
@@ -58,7 +59,7 @@ class UserRegistrationApiService {
         Log.info("Created organization. mdsId=$mdsId, userId=$userId.")
 
         if (firstUser) {
-            organizationManagementApiService.approveOrganization(mdsId, userId)
+            organizationRegistrationApiService.approveOrganization(mdsId, userId)
             Log.info("Approved organization for first user. mdsId=$mdsId, userId=$userId.")
         }