+++ authority-portal-frontend/src/app/core/utils/ui-utils.ts
@@ -4,11 +4,13 @@ export const getOrganizationRegistrationStatusClasses = (
   status: OrganizationRegistrationStatusDto,
 ): string => {
   switch (status) {
-    case OrganizationRegistrationStatusDto.Approved:
+    case OrganizationRegistrationStatusDto.Active:
       return 'text-emerald-700 bg-emerald-100/60';
     case OrganizationRegistrationStatusDto.Rejected:
       return 'text-red-700 bg-red-100/60';
     case OrganizationRegistrationStatusDto.Pending:
+      return 'bg-gray-100/90';
+    case OrganizationRegistrationStatusDto.Invited:
       return 'bg-gray-100/90';
     default:
       return '';